+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The following is a curated list of changes in the Enact agate module, newest cha
 
 ### Changed
 
+- `agate/ThemeDecorator` config `rootId` to specify React DOM tree root for global event handlers
 - `agate/ThemeDecorator` to exclude `enact-fit` className when `disableFullscreen` is true
 
 ## [1.1.2] - 2021-02-05
@@ -25,7 +26,6 @@ The following is a curated list of changes in the Enact agate module, newest cha
 - `agate/Input` prop `size` to change the size of both input and icons (default large)
 - `agate/Item` prop `centered`, `disabled`, `inline`, `marqueeOn`, and `size`
 - `agate/Popup` content padding and margin with `closeButton`
-- `agate/ThemeDecorator` config `rootId` to specify React DOM tree root for global event handlers
 - `agate/Popup` `bottom`, `fullscreen`, `left`, and `right` values to `position` prop
 - `agate/Popup` `duration` and `type` props