+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The following is a curated list of changes in the Enact agate module, newest cha
 - `agate/ContextualPopupDecorator` prop `css` to support customizations
 - `agate/Dropdown` prop `width` to support multiple widths
 - `agate/Item` to have a `selected` publicClassname
+- `agate/Spinner` prop `children` to support the display of text below the icon
 
 ### Changed