+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The following is a curated list of changes in the Enact agate module, newest cha
 ### Fixed
 
 - `agate/Button` style to match latest design for Silicon skin
+- `agate/SliderButton` 5-way navigation functionality to change value via direction keys
+- `agate/SliderButton` styling to visually show the focused state
 
 ## [1.1.2] - 2021-02-05