+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 The following is a curated list of changes in the Enact agate module, newest changes on the top.
 
-## [unreleased]
+## [2.0.0-alpha.1] - 2021-02-25
+
+-  The framework was updated to support React 17.0.1
 
 ### Added